[SUITE_STRUCTURE_STANDARD] passed.
[SUITE_PROCESS_STANDARD] start.
 move test related schema into S2GraphProvider from S2Graph.
